Fix 221 MYLIST handling, broken by the ANIME and EPISODE changes
The code in Client.pm that faked a 312 response depended on
data that the FILE query no longer returns, replaced by the ANIME
and EPISODe queries. Said code also ignored caching since Client.pm
has no knowledge of the cache.

Move the 221 handling to kadr.pl in order to use cached data, fix it
to use ANIME and EPISODE data.